Added error-checking for accountIDs. This should be replaced with actually passing const references to AccountInfo to database, but I'll do that later. CharacterID validation is still not done.

+++ gogo/src/modules/OnCharDelete.cpp
@@ -17,6 +17,9 @@ using namespace cockpit;
 
 void GoGoClient::OnCharDelete(MUID /* uidPlayer */, uint32_t charMarker, const std::string& /* charName */)
 {
+	if(!myAccount.isValid)
+		return transmitter->disconnect();
+
 	database->DeleteCharacter(myAccount.AccountId, charMarker);
 	transmitter->send(packet::protocol::Match_ResponseDeleteChar(PEC_NONE));
 }
